Minor change to avoid warning.

+++ src/main/java/ca/jeremycook/romannumeralconverter/RomanNumeralConverter.java
@@ -64,7 +64,7 @@ class RomanNumeralConverter
 
 			return boundariesKey != null ? (remainder - boundariesKey) : 0;
 		})
-				.limit(boundaries.size())
+				.limit((long) boundaries.size())
 				.filter(val -> val > 0)
 				.map(boundaries::floorKey)
 				.map(boundaries::get)
